“添加日志模块、pdf导出完善、考试页面添加简答题等，功能羁绊部分逻辑修复等”

+++ src/mvc/com/app/service/impl/ContestPaperService.java
@@ -11,11 +11,15 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
+import com.annotation.SystemServiceLog;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.app.dao.ContestMapper;
 import com.app.dao.ContestpaperMapper;
+import com.app.dao.SimproblemMapper;
 import com.code.model.Contestpaper;
+import com.code.model.Simproblem;
 import com.github.pagehelper.PageHelper;
 
 import net.sf.json.JSONArray;
@@ -25,6 +29,11 @@ public class ContestPaperService {
 
 	@Autowired
 	private ContestpaperMapper contestpaperDao;
+	@Autowired
+	private SimproblemMapper simpDao;
+	@Autowired
+	private ContestMapper contestDao;
+	
 	
 	public List GetAllContestPaper(String Keyword,String startTime,String endTime,int simCourseId,String pageSize,String pageNumber){
         //System.out.println("key=" + Keyword + " startTime=" + startTime + "endTime=" + endTime);
@@ -62,16 +71,25 @@ public class ContestPaperService {
 	    System.out.println(JSONArray.fromObject(list).toString());
 	    return list;
 	}
-	
+
+	@SystemServiceLog(description = "添加考试试卷")
 	public int AddContestPaper(Contestpaper contestpaper){
 		return contestpaperDao.insertSelective(contestpaper);
 	}
-	
+
+	@SystemServiceLog(description = "更新试卷信息")
 	public int UpdateContestPaper(Contestpaper contestpaper){
 		return contestpaperDao.updateByPrimaryKeySelective(contestpaper);
 	}
-	
+
+	@SystemServiceLog(description = "删除试卷")
 	public int DeleteContestPaper(String id){
+		//根据id查询试卷是否用作考试
+		List contestList = contestDao.selContestByPaperId(id);
+		if(!contestList.isEmpty()) {
+			return -2;
+		}
+		
 		return contestpaperDao.deleteByPrimaryKey(Integer.parseInt(id));
 	}
 	
@@ -81,8 +99,9 @@ public class ContestPaperService {
 		return contestpaperDao.selectByPrimaryKey(Integer.parseInt(id));
 	}
 	
-	public int DeleteAllContestPaper(List<String> ids)
-	{  
+	
+	@SystemServiceLog(description = "批量删除试卷")
+	public int DeleteAllContestPaper(List<String> ids) {  
 	   int count = 0;
 	   for(String id : ids){
 		 count += contestpaperDao.deleteByPrimaryKey(Integer.parseInt(id));
@@ -90,4 +109,32 @@ public class ContestPaperService {
 	   } 
 	   return count;
 	}
+	
+	/**
+	 * 查询可以添加复用题目的某课程下、某个教师出的试卷(排除已存在此题的试卷)(管理员出的试题都可以查询到)
+	 * @param teacherId
+	 * @param simCourseId
+	 * @return
+	 */
+	public List<Map<String,Object>> selReusePaper(String teacherId,int simCourseId) {
+		
+		List<Map<String,Object>> paperList = contestpaperDao.selReusePaper(teacherId,simCourseId);
+		
+		return paperList;
+		
+	}
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
 }
